版本更新为2.5.50： 1、预占申请单：解决当单据物料数目大于12条时，页面无法打开的问题 2、库存预占序时薄：不在显示预占数量为0的物料预占信息; 3、xx材料表:解决单据在未审核的状态下，无法修改设计人的问题; 4、解决领料到审核后，自动消除库存占用数量再某些情况下错误的问题，变量类型有double修改为decimal;

+++ code/code/borui/app/企业版/MainProgram/model/仓存管理/业务单据/MaterielProOccupiedDetails.cs
@@ -152,7 +152,7 @@ namespace MainProgram.model
             }
 
             ArrayList billNumList = MaterielProOccupiedOrder.getInctance().getAllReviewMaterielProBillNumFromProjectNum(projectNum);
-            double tmpValue = needCancelValue;
+            decimal tmpValue = (decimal)needCancelValue;
 
             for (int i = 0; i < billNumList.Count; i++)
             {
@@ -162,16 +162,16 @@ namespace MainProgram.model
                 {
                     if (index.Value.billNumber == billNumber && index.Value.materielID == materielID && index.Value.isCancel == "0")
                     {
-                        if (index.Value.value - tmpValue >= 0)
+                        if ((decimal)index.Value.value - tmpValue >= 0)
                         {
-                            updateRecordValue(index.Value.value - tmpValue, index.Value.pkey);
-                            tmpValue -= needCancelValue;
+                            updateRecordValue((decimal)index.Value.value - tmpValue, index.Value.pkey);
+                            tmpValue -= (decimal)needCancelValue;
                             break;
                         }
                         else
                         {
                             updateRecordValue(0, index.Value.pkey);
-                            tmpValue -= index.Value.value;
+                            tmpValue -= (decimal)index.Value.value;
                         }
                     }
                 }
@@ -186,7 +186,7 @@ namespace MainProgram.model
         }
 
         // 撤销单据
-        public void updateRecordValue(double value, int pkey)
+        public void updateRecordValue(decimal value, int pkey)
         {
             string delete = "UPDATE [dbo].[WAREHOUSE_MANAGEMENT_PRO_OCCUPIED_DETAILS] SET VALUE = " + value + " WHERE PKEY = " + pkey;
